# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RAFMARKET - Pi Network Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin/seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin/seller)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # CART
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, description, price, currency, stock, categoryId, sellerId, imageUrls, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        stock:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        sellerId:
          type: integer
        sellerUsername:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, sold_out]
        isFeatured:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name, description, price, currency, stock, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        stock:
          type: integer
        categoryId:
          type: integer
        imageUrls:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
    ProductPatch:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        stock:
          type: integer
        categoryId:
          type: integer
        imageUrls:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, sold_out]
        isFeatured:
          type: boolean
    ProductPage:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Cart
...