# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RAFMARKET - Pi Network Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  currency: string;
  stock: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  sellerId: number;
  /** @nullable */
  sellerUsername?: string | null;
  imageUrls: string[];
  tags?: string[];
  status: ProductStatus;
  isFeatured?: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  createdAt: Date;
}